throw error, reject promise and show alert if OCC worker fails with computation.

+++ projects/bitbybit/src/app/bitbybit-app/occ.worker.ts
@@ -28,45 +28,62 @@ addEventListener('message', ({ data }) => {
     const d = data as DataInput;
     let result;
 
-    // Ok, so this is baked in memoization as all OCC computations are potentially very expensive
-    // we can always return already computed entity hashes. On UI side we only deal with hashes as long
-    // as we don't need to render things and when we do need, we call tessellation methods with these hashes
-    // and receive real objects. This cache is useful in modeling operations throughout 'run' sessions.
-    if (d.action.functionName !== 'shapeToMesh' && d.action.functionName !== 'startedTheRun') {
-        // if inputs have shape or shapes properties, these are hashes on which the operations need to be performed.
-        // We thus replace these hashes to real objects from the cache before functions are called,
-        // this probably looks like smth generic but isn't, so will need to check if it works
-        if (d.action.inputs.shape) {
+    try {
+        // Ok, so this is baked in memoization as all OCC computations are potentially very expensive
+        // we can always return already computed entity hashes. On UI side we only deal with hashes as long
+        // as we don't need to render things and when we do need, we call tessellation methods with these hashes
+        // and receive real objects. This cache is useful in modeling operations throughout 'run' sessions.
+        if (d.action.functionName !== 'shapeToMesh' && d.action.functionName !== 'startedTheRun') {
+            // if inputs have shape or shapes properties, these are hashes on which the operations need to be performed.
+            // We thus replace these hashes to real objects from the cache before functions are called,
+            // this probably looks like smth generic but isn't, so will need to check if it works
+            if (d.action.inputs.shape) {
+                d.action.inputs.shape = cacheHelper.checkCache(d.action.inputs.shape);
+            }
+            if (d.action.inputs.shapes && d.action.inputs.shapes.length > 0) {
+                d.action.inputs.shapes = d.action.inputs.shapes.map(hash => cacheHelper.checkCache(hash));
+            }
+
+            result = cacheHelper.cacheOp(d.action, () => openCascade[d.action.functionName](d.action.inputs)).hash;
+        }
+        if (d.action.functionName === 'shapeToMesh') {
             d.action.inputs.shape = cacheHelper.checkCache(d.action.inputs.shape);
+            result = openCascade[d.action.functionName](d.action.inputs.shape, d.action.inputs.precision);
         }
-        if (d.action.inputs.shapes && d.action.inputs.shapes.length > 0) {
-            d.action.inputs.shapes = d.action.inputs.shapes.map(hash => cacheHelper.checkCache(hash));
+        // Only the cache that was created in previous run has to be kept, the rest needs to go
+        if (d.action.functionName === 'startedTheRun') {
+            if (Object.keys(cacheHelper.hashesFromPreviousRun).length > 0) {
+                cacheHelper.cleanUpCache();
+            }
+            result = {
+                argCache: Object.keys(cacheHelper.argCache),
+                hashesFromPreviousRun: Object.keys(cacheHelper.hashesFromPreviousRun),
+                usedHashes: Object.keys(cacheHelper.usedHashes),
+            };
         }
+        // Returns only the hash as main process can't receive pointers
+        // But with hash reference we can always initiate further computations
+        postMessage({
+            uid: data.uid,
+            result
+        });
+    } catch (e) {
+        let props;
+        if(d && d.action && d.action.inputs){
+            props = `Input values were: {${Object.keys(d.action.inputs).map(key => `${key}: ${d.action.inputs[key]}`).join(',')}}. `;
+        }
+        let fun;
+        if(d && d.action && d.action.functionName){
+            fun = `- ${d.action.functionName}`;
+        }
+        const message = 
+        postMessage({
+            uid: data.uid,
+            result: undefined,
+            error: `OCC computation failed when executing function ${fun}. ${props}Original message: ${e}`
+        });
+    }
 
-        result = cacheHelper.cacheOp(d.action, () => openCascade[d.action.functionName](d.action.inputs)).hash;
-    }
-    if (d.action.functionName === 'shapeToMesh') {
-        d.action.inputs.shape = cacheHelper.checkCache(d.action.inputs.shape);
-        result = openCascade[d.action.functionName](d.action.inputs.shape, d.action.inputs.precision);
-    }
-    // Only the cache that was created in previous run has to be kept, the rest needs to go
-    if (d.action.functionName === 'startedTheRun') {
-        if (Object.keys(cacheHelper.hashesFromPreviousRun).length > 0) {
-            cacheHelper.cleanUpCache();
-        }
-        result = {
-            argCache: Object.keys(cacheHelper.argCache),
-            hashesFromPreviousRun: Object.keys(cacheHelper.hashesFromPreviousRun),
-            usedHashes: Object.keys(cacheHelper.usedHashes),
-        };
-    }
-
-    // Returns only the hash as main process can't receive pointers
-    // But with hash reference we can always initiate further computations
-    postMessage({
-        uid: data.uid,
-        result
-    });
 });
 
 class OccHelper {
@@ -598,9 +615,6 @@ export class Occ {
                             thisFace.vertex_coord[((vertexIndex - 1) * 3) + 1],
                             thisFace.vertex_coord[((vertexIndex - 1) * 3) + 2]
                         ]);
-                        // thisEdge.vertex_coord[(j * 3) + 0] = thisFace.vertex_coord[((vertexIndex - 1) * 3) + 0];
-                        // thisEdge.vertex_coord[(j * 3) + 1] = thisFace.vertex_coord[((vertexIndex - 1) * 3) + 1];
-                        // thisEdge.vertex_coord[(j * 3) + 2] = thisFace.vertex_coord[((vertexIndex - 1) * 3) + 2];
                     }
 
                     thisEdge.edge_index = fullShapeEdgeHashes[edgeHash];
@@ -758,8 +772,10 @@ class CacheHelper {
 
         usedHashKeys.forEach(hash => {
             if (!hashesFromPreviousRunKeys.find(h => h === hash)) {
-                this.argCache[hash].delete();
-                delete this.argCache[hash];
+                if (this.argCache[hash]) {
+                    this.argCache[hash].delete();
+                    delete this.argCache[hash];
+                }
                 delete this.usedHashes[hash];
             }
         });
